% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.r
\name{packages}
\alias{packages}
\alias{find_packages}
\alias{installed_packages}
\alias{package_dependencies}
\title{Local package paths and dependencies}
\usage{
find_packages(x)

installed_packages(x)

package_dependencies(x, which = "strong")
}
\arguments{
\item{x}{for `find_packages()`, `installed_packages()` a character vector of package names,
for `package_dependencies()` a character vector of package names, local directories,
or a path to tarballs.}

\item{which}{for `package_dependencies()` which dependencies to return, see details}
}
\value{
a named logical vector
}
\description{
Find locally installed packages and their dependencies
}
\details{
These functions replace/extend similar functions in base and utils,
but provide slightly different interface and/or functionality.

`find_packages()` explores `.libPaths()` to find locally installed packages
and returns their path, similar to [base::find.package()].
It is less chatty than the base function and returns `NA` when the package was not found,
instead of omitting the value altogether.
`find_packages()` returns a character vector of the same length as input with paths
towards the packages, and `NA` if the package wasn't found.

`installed_packages()` is simple wrapper around `find_packages()` to check
whether the specified packages are installed. Compared to [utils::installed.packages()],
it doesn't return a list of all installed packages, making it much faster than
`pkg %in% rownames(utils::installed.packages())`.
It returns a logical vecor of the same length as input with `TRUE` if the package is found
in `.libPaths()` and `FALSE` otherwise.

`package_dependencies()` will return dependencies of a locally installed package,
a package in a local directory, or a tarball (as obtained from `R CMD build`).
This is done by parsing the DESCRIPTION file and extracting the `Depends`, `Imports`,
`LinkingTo`, and `Enhances` fields.
Unlike [tools::package_dependencies()], it works only with locally available packages
and it doesn't explore CRAN or require internet access, but works on packages
that are not available on CRAN..
`package_dependencies()` returns a list with the same name as input, with each element
being a character vector of package dependencies.

The `which` argument for `package_dependencies()` can be one or more of the dependency types
(`Depends`, `Imports`, `LinkingTo`, `Enhances`), or `strong` for the first three types,
`most` for the first four types, or `all` for all dependency types.
}
\examples{
find_packages(c("base", "mpd", "not even valid name"))
installed_packages(c("base", "mpd", "not even valid name"))

package_dependencies(c("base", "mpd"), which = "all")

}
\seealso{
[utils::installed.packages()] for a database of all installed packages,
[tools::package_dependencies()] for remote the 
[base::system.file()] and [base::find.package()] to get file/folder within package directory
[base::.libPaths()] for a vector of all library directories
}
