% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.r
\name{test}
\alias{test}
\alias{check}
\alias{document}
\alias{install}
\title{Minimal package development functions}
\usage{
test(pkg = ".", reset_lib = TRUE)

check(pkg = ".", as_cran = FALSE)

document(pkg = ".")

install(pkg = ".")
}
\arguments{
\item{pkg}{a path to the package directory}

\item{reset_lib}{reset the search path after exit (default: TRUE)}

\item{as_cran}{run the `R CMD check` with the CRAN preset (internet connection required)}
}
\value{
nothing, these functions are run for their side-effect
}
\description{
These functions provide a dependency-light alternative to `devtools`.
}
\details{
`test()` install the package and runs tests in the `test` folder, these tests must be created by
some unit-testing framework in that they need to be able to be run outside of `R CMD check`
in the sense that they must report an error if the test fail, as the `Rout.save` snapshot
testing isn't supported when you just run the files in the `test` folder.

The `check()` maps to the `R CMD check`.

The `document()` runs the `roxygen2::roxygenize` function to generate documentation.

The `install()` installs the package locally.
}
